Walls Ready , EventListners - Keyboard For Movement
Now the pacman is moving free through the map;
Debug=
there was a parathisis mmissing in one of the Collision parameter making our little pacman go into the walls.

there was also an opening in right side of the map that was letting our pacman leave the route,
Fix for this was rather simple,
I just made some changes in the Map{}
array and the map was closed.